Re-enable results testing (#159)
Re-enable results testing

Commit 0d1e57c removed "result testing" from the CI, by not including test/
result.py in pytest. This was because PyTorch is not deterministic and cannot be
made so.
Re-enable it, and hide PyTorch hash testing behind a toggle set to False.
